Refactor whisper transcription services
