+++ appsettings.json
@@ -14,7 +14,15 @@
   "Whisper": {
     "ExecutablePath": "whisper",
     "Model": "medium",
-    "UseGpu": true
+    "UseGpu": true,
+    "Provider": "whisper"
+  },
+
+  "FasterWhisper": {
+    "ExecutablePath": "faster-whisper",
+    "Model": "medium",
+    "Device": "cuda",
+    "ComputeType": "float16"
   },
 
   "FfmpegExePath": "C:\\Documents and Settings\\user\\AppData\\Local\\Programs\\FFmpeg\\bin\\",
